requirements.txt
# Core dependencies for NeRF Few-Shot Analysis
# This file contains all dependencies needed to reproduce the experiments

# PyTorch ecosystem
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Vision and ML libraries
transformers>=4.20.0
timm>=0.6.0
opencv-python>=4.5.0
Pillow>=8.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Image processing
imageio>=2.9.0
imageio-ffmpeg>=0.4.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.62.0
tensorboard>=2.8.0
wandb>=0.12.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0

# Optional: CUDA support (uncomment if using NVIDIA GPU)
# torch>=2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision>=0.15.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
